Updated the testcases
Updated the testcases
