+++ src/main/java/pages/MedianAuthRole.java
@@ -7,13 +7,8 @@ import base.MedianSpecificMethods;
 
 public class MedianAuthRole extends MedianSpecificMethods 
 {
-	public MedianAuthRole selectRoleAuth() {
-		clickElement(locateElement("xpath", props.getProperty("Median.SystemMaintenance.RoleName.Xpath")));
-		wait(5000);
-		return this;
-	}
 	
-	
+
 	public MedianAuthRole clickAuthButton() {
 		wait(2000);
 		clickElement(locateElement("xpath",props.getProperty("Median.SystemMaintenance.Authenticate.Role.Xpath")));
