+++ src/main/java/testcases/MedianEditExternalSystemTC01.java
@@ -13,8 +13,9 @@ public class MedianEditExternalSystemTC01 extends MedianSpecificMethods {
 		login.enterMakerUsername()
 		.enterMakerPassword()
 		.clickSignInButton()
-		.fileUploadMaintenance()
-		.selectNew()
+		.clickfileUploadMaintenance()
+		.clickexternalSystem()
+		.selectEdit()
 		.enterExternalSystemCode()
 		.enterExternalSystemName()
 		.selectDepartment()
