+++ src/main/java/testcases/MedianNewExternalSystemTC01.java
@@ -13,7 +13,8 @@ public class MedianNewExternalSystemTC01 extends MedianSpecificMethods {
 		login.enterMakerUsername()
 		.enterMakerPassword()
 		.clickSignInButton()
-		.fileUploadMaintenance()
+		.clickfileUploadMaintenance()
+		.clickexternalSystem()
 		.selectNew()
 		.enterExternalSystemCode()
 		.enterExternalSystemName()
